pyproject.toml
[project]
name = "camlin-backend-test"
version = "0.1.0"
description = "Camlin take home test for Senior Backend Engineer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.13",
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
    "slowapi>=0.1.9",
    "uvicorn>=0.34.3",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[project.optional-dependencies]
dev = [
  "pytest",
]
